feat(es_recommend_optimize): Combined code into one file for maintainability
